+++ src/main/java/memethespire/ConfigMenu.java
@@ -1,19 +1,34 @@
 package memethespire;
 
+import basemod.ModButton;
+import basemod.ModLabel;
 import basemod.ModLabeledToggleButton;
 import basemod.ModPanel;
 import com.megacrit.cardcrawl.core.Settings;
 import com.megacrit.cardcrawl.helpers.FontHelper;
 
 public class ConfigMenu extends ModPanel {
-    private static float TOGGLE_BUTTON_X = 360.0f;
-    private static float TOGGLE_BUTTON_Y = 750.0f;
+    private static float UI_X = 360.0f;
+    private static float UI_Y = 650.0f;
     private static float TOGGLE_BUTTON_SPACE = 70.0f;
 
     public ConfigMenu() {
+        ModLabel buttonLabel = new ModLabel(
+                "Reload Meme Collections",
+                UI_X + 120,
+                UI_Y + 50, this, (me) -> {});
+        this.addUIElement(buttonLabel);
+
+        ModButton consoleKeyButton = new ModButton(UI_X, UI_Y, this,
+                (me) -> {
+            MemeCollection.collections.clear();
+            MemeCollection.loadMemeCollections();
+        });
+        this.addUIElement(consoleKeyButton);
+
         ModLabeledToggleButton enableCardModifications = new ModLabeledToggleButton(
                 "Enable Card Modifications",
-                TOGGLE_BUTTON_X, TOGGLE_BUTTON_Y - TOGGLE_BUTTON_SPACE * 0,
+                UI_X, UI_Y - TOGGLE_BUTTON_SPACE * 1,
                 Settings.CREAM_COLOR, FontHelper.charDescFont,
                 MemeTheSpire.config.enableCardModifications,
                 this, (modLabel) -> {}, (modToggleButton) -> {
@@ -24,7 +39,7 @@ public class ConfigMenu extends ModPanel {
 
         ModLabeledToggleButton enableTooltipMemes = new ModLabeledToggleButton(
                 "Enable Tooltip Memes",
-                TOGGLE_BUTTON_X, TOGGLE_BUTTON_Y - TOGGLE_BUTTON_SPACE * 1,
+                UI_X, UI_Y - TOGGLE_BUTTON_SPACE * 2,
                 Settings.CREAM_COLOR, FontHelper.charDescFont,
                 MemeTheSpire.config.enableTooltipMemes,
                 this, (modLabel) -> {}, (modToggleButton) -> {
@@ -35,7 +50,7 @@ public class ConfigMenu extends ModPanel {
 
         ModLabeledToggleButton enableCardPlayMessages = new ModLabeledToggleButton(
                 "Enable Card Play Messages",
-                TOGGLE_BUTTON_X, TOGGLE_BUTTON_Y - TOGGLE_BUTTON_SPACE * 2,
+                UI_X, UI_Y - TOGGLE_BUTTON_SPACE * 3,
                 Settings.CREAM_COLOR, FontHelper.charDescFont,
                 MemeTheSpire.config.enableCardPlayMessages,
                 this, (modLabel) -> {}, (modToggleButton) -> {
@@ -46,7 +61,7 @@ public class ConfigMenu extends ModPanel {
 
         ModLabeledToggleButton enableRelicModifications = new ModLabeledToggleButton(
                 "Enable Relic Modifications",
-                TOGGLE_BUTTON_X, TOGGLE_BUTTON_Y - TOGGLE_BUTTON_SPACE * 3,
+                UI_X, UI_Y - TOGGLE_BUTTON_SPACE * 4,
                 Settings.CREAM_COLOR, FontHelper.charDescFont,
                 MemeTheSpire.config.enableRelicModifications,
                 this, (modLabel) -> {}, (modToggleButton) -> {
@@ -57,7 +72,7 @@ public class ConfigMenu extends ModPanel {
 
         ModLabeledToggleButton enableEnemyDialogues = new ModLabeledToggleButton(
                 "Enable Enemy Dialogues",
-                TOGGLE_BUTTON_X, TOGGLE_BUTTON_Y - TOGGLE_BUTTON_SPACE * 4,
+                UI_X, UI_Y - TOGGLE_BUTTON_SPACE * 5,
                 Settings.CREAM_COLOR, FontHelper.charDescFont,
                 MemeTheSpire.config.enableEnemyDialogues,
                 this, (modLabel) -> {}, (modToggleButton) -> {